+++ src/features/suggestions/Unit.ts
@@ -71,3 +71,13 @@ export function isConvertible(fromUnit: Unit, toUnit: Unit) {
     .possibilities()
     .includes(toUnit);
 }
+
+export function isMeasurementConvertibleFunc(fromUnit: Unit) {
+  return (measurement: { unit?: Unit; }) => {
+    const { unit: toUnit } = measurement;
+    if (_.isUndefined(toUnit))
+      return false;
+
+    return isConvertible(fromUnit, toUnit);
+  };
+}
